MuonMatcher::matchSimple()
added method MuonMatcher::matchSimple(), which does no propagation, but
only checks the delta phi and eta between the vertex and OMTF candidate.
It is useful for the prompt muons in singleMu samples, as the
propagation failes for the muons with pT < ~3.2 GeV

L1MuonAnalyzerOmtf.cc
setting up the eta range for the PtGenVsPtCand analyzer depending on
whether the propagation is used or not. Without propagation, the cut on
the vertex eta in the OMTF region is applied with used of the function
simTrackIsMuonInOmtf. With propagation, the propagated eta at the second
muons station is checked.
For simTracks the muonMatcher.matchSimple if matchUsingPropagation is
false.